+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Breaking changes:
 * [OLMIS-3586](https://openlmis.atlassian.net/browse/OLMIS-3586): Changed `GET /api/facilities` and `POST /api/facilities/search` endpoints to do conjunction search.
 * [OLMIS-5000](https://openlmis.atlassian.net/browse/OLMIS-5000): Merged all supply line searching endpoints to on GET /api/supplyLines endpoint with programId, supervisoryNodeId and supplyingFacilityId (multiple) parameters.
 * [OLMIS-5853](https://openlmis.atlassian.net/browse/OLMIS-5853):change supervisory node edit validation sequence fix functional test error
+
 New functionality added in a backwards-compatible manner:
 * [OLMIS-5990](https://openlmis.atlassian.net/browse/OLMIS-5990): Added ability to search roles by right ID
 * [OLMIS-683](https://openlmis.atlassian.net/browse/OLMIS-683): Update orderable model to Allow for definition of kits.
@@ -18,6 +19,10 @@ Improvements:
 * [OLMIS-5853](https://openlmis.atlassian.net/browse/OLMIS-5853): Add Unique Name constraint on Supervisory Node
 * [OLMIS-5852](https://openlmis.atlassian.net/browse/OLMIS-5852): Add Case Insensitive Unique Code constraint on Supervisory Node
 * [OLMIS-6004](https://openlmis.atlassian.net/browse/OLMIS-6004) : Add sortable parameter to program search api
+
+Bug fixes:
+* [OLMIS-5981](https://openlmis.atlassian.net/browse/OLMIS-5981): Updated /supervisoryNodes/{supervisoryNode_id} PUT to throw exception when requisition group is missing.
+
 12.0.0 / 2018-12-12
 ===================